moved pricing definitions list to pricing tab (instead of having a second tab) oonly visible for slf tenant for prod  testing
