+++ src/app/pages/settings-integration/pricing/pricing-definitions/settings-pricing-definitions.component.html
@@ -1,0 +1,12 @@
+<div class="card card-big">
+  <div class="card-header card-header-icon card-header-primary">
+    <div class="card-icon">
+      <mat-icon>money</mat-icon>
+    </div>
+    <h3 class="card-title">{{'settings.pricing.title' | translate}}</h3>
+  </div>
+  <div class="card-body pricing-card-body">
+    <app-table [dataSource]="settingsPricingsTableDataSource"></app-table>
+  </div>
+</div>
+
