+++ src/app/pages/settings-integration/pricing/pricing-definitions/settings-pricing-definitions.component.ts
@@ -1,0 +1,24 @@
+import { Component } from '@angular/core';
+import { FormGroup } from '@angular/forms';
+
+import { ComponentService } from '../../../../services/component.service';
+import { AnalyticsSettings } from '../../../../types/Setting';
+import { TenantComponents } from '../../../../types/Tenant';
+import { SettingsPricingDefinitionsTableDataSource } from './settings-pricing-definition-table-data-source';
+
+@Component({
+  selector: 'app-settings-pricing-definitions',
+  templateUrl: 'settings-pricing-definitions.component.html',
+})
+export class SettingsPricingDefinitionsComponent {
+  public isActive = false;
+  public analyticsSettings!: AnalyticsSettings;
+  public formGroup: FormGroup;
+
+  public constructor(
+    componentService: ComponentService,
+    public settingsPricingsTableDataSource: SettingsPricingDefinitionsTableDataSource,
+  ) {
+    this.isActive = componentService.isActive(TenantComponents.PRICING);
+  }
+}
