+++ src/app/pages/settings-integration/pricing/settings-pricing.component.html
@@ -22,4 +22,6 @@
   <app-settings-simple-pricing *ngIf="isActive && pricingSettings?.type === 'simple'" [pricingSettings]="pricingSettings" [formGroup]="formGroup"></app-settings-simple-pricing>
   <!-- Convergent Charging Pricing -->
   <app-settings-convergent-charging *ngIf="isActive && pricingSettings?.type === 'convergentCharging'" [pricingSettings]="pricingSettings" [formGroup]="formGroup"></app-settings-convergent-charging>
+  <!-- Pricing Definitions -->
+  <app-settings-pricing-definitions *ngIf="isActive && pricingSettings?.type === 'simple' && isTestTenant"></app-settings-pricing-definitions>
 </form>
