--- src/app/pages/settings-integration/pricings/settings-pricings.component.html
@@ -1,27 +1,0 @@
-<div *ngIf="!isActive" class="card">
-  <div class="card-header card-header-icon card-header-primary">
-    <div class="card-icon">
-      <mat-icon>info</mat-icon>
-    </div>
-    <h3 class="card-title">{{'settings.information' | translate}}</h3>
-  </div>
-  <div class="card-body">
-    <span [innerHTML]="'settings.pricing.full_description' | translate"></span>
-    <a href="https://help.sap.com/viewer/product/SAP_CONVERGENT_CHARGING/4.0/en-US" target="new"><u>SAP Convergent Charging</u></a><br><br>
-    {{'settings.activation_contact_msg' | translate}}
-  </div>
-</div>
-
-<div *ngIf="isActive">
-  <div class="card card-big">
-    <div class="card-header card-header-icon card-header-primary">
-      <div class="card-icon">
-        <mat-icon>money</mat-icon>
-      </div>
-      <h3 class="card-title">{{'settings.pricing.title' | translate}}</h3>
-    </div>
-    <div class="card-body pricing-card-body">
-      <app-table [dataSource]="settingsPricingsTableDataSource"></app-table>
-    </div>
-  </div>
-</div>
