+++ src/app/pages/settings-integration/settings-integration.component.html
@@ -38,15 +38,6 @@
       </ng-template>
     </mat-tab>
     
-    <mat-tab *ngIf="isTestTenant">
-      <ng-template mat-tab-label>
-        <mat-icon>money</mat-icon><span>{{'settings.tabs.pricings' | translate}}</span>
-      </ng-template>
-      <ng-template matTabContent>
-        <app-settings-pricings></app-settings-pricings>
-      </ng-template>
-    </mat-tab>
-
     <mat-tab>
       <ng-template mat-tab-label>
         <mat-icon>payment</mat-icon><span>{{'settings.tabs.billing' | translate}}</span>
