+++ src/assets/scss/views/pricing/pricing.scss
@@ -17,7 +17,7 @@
 }
 
 .pricing-card-body {
-  height: 80vh;
+  height: 60vh;
 }
 
 .dimension-toggle {
